调试修改

+++ dos/dos-imp/src/main/java/allthings/iot/dos/service/IotTagServiceImpl.java
@@ -1,15 +1,17 @@
 package allthings.iot.dos.service;
 
 import allthings.iot.common.dto.ResultDTO;
+import allthings.iot.dos.api.IotDeviceTypeService;
+import allthings.iot.dos.api.IotTagService;
 import allthings.iot.dos.constant.ErrorCode;
 import allthings.iot.dos.dao.IotDeviceTagDao;
 import allthings.iot.dos.dao.IotTagDao;
 import allthings.iot.dos.dto.IotDeviceDTO;
+import allthings.iot.dos.dto.IotDeviceTagDTO;
 import allthings.iot.dos.dto.IotTagDTO;
 import allthings.iot.dos.dto.query.IotTagQueryDTO;
 import allthings.iot.dos.model.IotDeviceTag;
 import allthings.iot.dos.model.IotTag;
-import allthings.iot.dos.api.IotTagService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -42,17 +44,36 @@ public class IotTagServiceImpl implements IotTagService {
 
     @Autowired
     private IotDeviceTagDao iotDeviceTagDao;
+    @Autowired
+    private IotDeviceTypeService iotDeviceTypeService;
 
     @Override
-    public ResultDTO<List<IotTagQueryDTO>> getIotTagList(Long iotProjectId) {
+    public ResultDTO<List<IotTagQueryDTO>> getIotTagList(IotTagDTO iotTagDTO) {
+
+        ResultDTO<Integer> judge = iotDeviceTypeService.judgeProject(iotTagDTO.getIotProjectId(),
+                iotTagDTO.getCreateOperatorId(), iotTagDTO.getRoleCode());
+        if (!judge.isSuccess()) {
+            return ResultDTO.newFail(ErrorCode.ERROR_8014.getMessage());
+        }
+
+        Long iotProjectId = iotTagDTO.getIotProjectId();
         List<IotTagQueryDTO> list = iotTagDao.getAllByIsDeleted(false, iotProjectId);
         return ResultDTO.newSuccess(list);
     }
 
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     @Override
-    public ResultDTO<List<IotDeviceDTO>> getDeviceByIotTagIdAndIotProjectId(Long iotTagId, Long iotProjectId,
-                                                                            String keywords, Boolean choose) {
+    public ResultDTO<List<IotDeviceDTO>> getDeviceByIotTagIdAndIotProjectId(IotTagQueryDTO tagQueryDTO, Boolean choose) {
+
+        ResultDTO<Integer> judge = iotDeviceTypeService.judgeProject(tagQueryDTO.getIotProjectId(),
+                tagQueryDTO.getCreateOperatorId(), tagQueryDTO.getRoleCode());
+        if (!judge.isSuccess()) {
+            return ResultDTO.newFail(ErrorCode.ERROR_8014.getMessage());
+        }
+
+        Long iotTagId = tagQueryDTO.getIotTagId();
+        Long iotProjectId = tagQueryDTO.getIotProjectId();
+        String keywords = tagQueryDTO.getKeywords();
         if (StringUtils.isEmpty(keywords)) {
             keywords = "";
         }
@@ -80,7 +101,17 @@ public class IotTagServiceImpl implements IotTagService {
 
     @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
     @Override
-    public ResultDTO<Integer> deleteTagByTagId(Long iotTagId, Long iotProjectId, Long operatorId) {
+    public ResultDTO<Integer> deleteTagByTagId(IotTagDTO iotTagDTO) {
+
+        ResultDTO<Integer> judge = iotDeviceTypeService.judgeProject(iotTagDTO.getIotProjectId(),
+                iotTagDTO.getCreateOperatorId(), iotTagDTO.getRoleCode());
+        if (!judge.isSuccess()) {
+            return ResultDTO.newFail(ErrorCode.ERROR_8014.getMessage());
+        }
+
+        Long iotTagId = iotTagDTO.getIotTagId();
+        Long iotProjectId = iotTagDTO.getIotProjectId();
+        Long operatorId = iotTagDTO.getCreateOperatorId();
         iotTagDao.deleteAllDeviceByTagId(iotTagId, operatorId);
         iotTagDao.deleteTagByTagId(iotTagId, iotProjectId, operatorId);
         return ResultDTO.newSuccess();
@@ -88,7 +119,14 @@ public class IotTagServiceImpl implements IotTagService {
 
     @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
     @Override
-    public ResultDTO<Integer> saveTag(IotTagDTO iotTagDto, Long operatorId) {
+    public ResultDTO<Integer> saveTag(IotTagDTO iotTagDto) {
+
+        ResultDTO<Integer> judge = iotDeviceTypeService.judgeProject(iotTagDto.getIotProjectId(),
+                iotTagDto.getCreateOperatorId(), iotTagDto.getRoleCode());
+        if (!judge.isSuccess()) {
+            return ResultDTO.newFail(ErrorCode.ERROR_8014.getMessage());
+        }
+
         int maxNameLen = 32;
         if (StringUtils.isEmpty(iotTagDto.getTagName()) || iotTagDto.getTagName().trim().length() > maxNameLen) {
             return ResultDTO.newFail(ErrorCode.ERROR_9001.getMessage());
@@ -99,23 +137,31 @@ public class IotTagServiceImpl implements IotTagService {
         }
         IotTag iotTag = new IotTag();
         BeanUtils.copyProperties(iotTagDto, iotTag);
-        iotTag.setCreateOperatorId(operatorId);
+        iotTag.setCreateOperatorId(iotTagDto.getCreateOperatorId());
         iotTagDao.saveAndFlush(iotTag);
         return ResultDTO.newSuccess();
     }
 
     @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
     @Override
-    public ResultDTO<Integer> saveDeviceOfTag(List<Long> iotDeviceIds, Long iotTagId, Long operatorId) {
+    public ResultDTO<Integer> saveDeviceOfTag(IotDeviceTagDTO deviceTagDTO) {
+
+        ResultDTO<Integer> judge = iotDeviceTypeService.judgeProject(deviceTagDTO.getIotProjectId(),
+                deviceTagDTO.getCreateOperatorId(), deviceTagDTO.getRoleCode());
+        if (!judge.isSuccess()) {
+            return ResultDTO.newFail(ErrorCode.ERROR_8014.getMessage());
+        }
+
+        List<Long> iotDeviceIds = deviceTagDTO.getIotDeviceIds();
         if (CollectionUtils.isEmpty(iotDeviceIds)) {
             return ResultDTO.newSuccess();
         }
         List<IotDeviceTag> iotDeviceTags = new ArrayList<>();
         for (Long iotDeviceId : iotDeviceIds) {
             IotDeviceTag iotDeviceTag = new IotDeviceTag();
-            iotDeviceTag.setCreateOperatorId(operatorId);
+            iotDeviceTag.setCreateOperatorId(deviceTagDTO.getCreateOperatorId());
             iotDeviceTag.setIotDeviceId(iotDeviceId);
-            iotDeviceTag.setIotTagId(iotTagId);
+            iotDeviceTag.setIotTagId(deviceTagDTO.getIotTagId());
             iotDeviceTags.add(iotDeviceTag);
         }
         iotDeviceTagDao.saveAll(iotDeviceTags);
@@ -124,11 +170,18 @@ public class IotTagServiceImpl implements IotTagService {
 
     @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
     @Override
-    public ResultDTO<Integer> deleteDeviceOfTag(List<Long> iotDeviceIds, Long iotTagId, Long operatorId) {
-        if (CollectionUtils.isEmpty(iotDeviceIds)) {
+    public ResultDTO<Integer> deleteDeviceOfTag(IotDeviceTagDTO deviceTagDTO) {
+
+        ResultDTO<Integer> judge = iotDeviceTypeService.judgeProject(deviceTagDTO.getIotProjectId(),
+                deviceTagDTO.getCreateOperatorId(), deviceTagDTO.getRoleCode());
+        if (!judge.isSuccess()) {
+            return ResultDTO.newFail(ErrorCode.ERROR_8014.getMessage());
+        }
+
+        if (CollectionUtils.isEmpty(deviceTagDTO.getIotDeviceIds())) {
             return ResultDTO.newSuccess();
         }
-        iotTagDao.deleteDeviceOfTag(iotDeviceIds, iotTagId, operatorId);
+        iotTagDao.deleteDeviceOfTag(deviceTagDTO.getIotDeviceIds(), deviceTagDTO.getIotTagId(), deviceTagDTO.getCreateOperatorId());
         return ResultDTO.newSuccess();
     }
